Cobbling a settings manager together
Creating settings manager, will be able to create settings profiles that you can apply to game. SettingsManager has a bunch of classes to help managed settings and profiles these profiles can then be Apply()'d (not tested lol). Currently have a WIP editor to create profiles, next to implement is the ability to add settings to the profiles and allow to configure a value for that setting.

Dummy code for "Conditionals" work, leaving in here as a reminder what to not do, going to research C# tech with RTTI to hopefully make something much better

Partial refactor of my jokebase

Updated plugin to work with Dalamud 11.0.1